// models/get-exchange-rate200-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Twelve Data API client for Node.js
 *
 * NOTE: This code is auto generated, please do not edit it manually.
 */

export interface GetExchangeRate200Response {
  /**
   * Requested currency symbol
   */
  symbol?: string;
  /**
   * Real-time exchange rate for the corresponding symbol
   */
  rate?: number;
  /**
   * Unix timestamp of the rate
   */
  timestamp?: number;
}